+++ src/ptex/PtexTriangleFilter.cpp
@@ -42,7 +42,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
 #include "PtexUtils.h"
 
 namespace {
-    inline double squared(double x) { return x*x; }
+    inline float squared(float x) { return x*x; }
 }
 
 void PtexTriangleFilter::eval(float* result, int firstChan, int nChannels,
@@ -82,16 +82,16 @@ void PtexTriangleFilter::eval(float* result, int firstChan, int nChannels,
     // accumulate the weight as we apply
     _weight = 0;
 
-    // allocate temporary double-precision result
-    _result = (double*) alloca(sizeof(double)*_nchan);
-    memset(_result, 0, sizeof(double)*_nchan);
+    // allocate temporary result
+    _result = (float*) alloca(sizeof(float)*_nchan);
+    memset(_result, 0, sizeof(float)*_nchan);
 
     // apply to faces
     splitAndApply(k, faceid, f);
 
     // normalize (both for data type and cumulative kernel weight applied)
     // and output result
-    double scale = 1.0 / (_weight * OneValue(_dt));
+    float scale = 1.0f / (_weight * OneValue(_dt));
     for (int i = 0; i < _nchan; i++) result[i] = float(_result[i] * scale);
 
     // clear temp result
@@ -104,60 +104,60 @@ void PtexTriangleFilter::buildKernel(PtexTriangleKernel& k, float u, float v,
 				     float uw1, float vw1, float uw2, float vw2,
 				     float width, float blur, Res faceRes)
 {
-    const double sqrt3 = 1.7320508075688772;
+    const float sqrt3 = 1.7320508075688772;
 
     // compute ellipse coefficients, A*u^2 + B*u*v + C*v^2 == AC - B^2/4
-    double scaleAC = 0.25 * width*width;
-    double scaleB = -2 * scaleAC;
-    double A = (vw1*vw1 + vw2*vw2) * scaleAC;
-    double B = (uw1*vw1 + uw2*vw2) * scaleB;
-    double C = (uw1*uw1 + uw2*uw2) * scaleAC;
+    float scaleAC = 0.25f * width*width;
+    float scaleB = -2.0f * scaleAC;
+    float A = (vw1*vw1 + vw2*vw2) * scaleAC;
+    float B = (uw1*vw1 + uw2*vw2) * scaleB;
+    float C = (uw1*uw1 + uw2*uw2) * scaleAC;
 
     // convert to cartesian domain
-    double Ac = 0.75 * A;
-    double Bc = (sqrt3/2) * (B-A);
-    double Cc = 0.25 * A - 0.5 * B + C;
+    float Ac = 0.75f * A;
+    float Bc = float(sqrt3/2) * (B-A);
+    float Cc = 0.25f * A - 0.5f * B + C;
 
     // compute min blur for eccentricity clamping
-    const double maxEcc = 15; // max eccentricity
-    const double eccRatio = (maxEcc*maxEcc + 1) / (maxEcc*maxEcc - 1);
-    double X = sqrt(squared(Ac - Cc) + squared(Bc));
-    double b_e = 0.5 * (eccRatio * X - (Ac + Cc));
+    const float maxEcc = 15.0f; // max eccentricity
+    const float eccRatio = (maxEcc*maxEcc + 1.0f) / (maxEcc*maxEcc - 1.0f);
+    float X = sqrtf(squared(Ac - Cc) + squared(Bc));
+    float b_e = 0.5f * (eccRatio * X - (Ac + Cc));
 
     // compute min blur for texel clamping
     // (ensure that ellipse is no smaller than a texel)
-    double b_t = squared(0.5 / faceRes.u());
+    float b_t = squared(0.5f / faceRes.u());
 
     // add blur
-    double b_b = 0.25 * blur * blur;
-    double b = PtexUtils::max(b_b, PtexUtils::max(b_e, b_t));
+    float b_b = 0.25f * blur * blur;
+    float b = PtexUtils::max(b_b, PtexUtils::max(b_e, b_t));
     Ac += b;
     Cc += b;
 
     // compute minor radius
-    double m = sqrt(2*(Ac*Cc - 0.25*Bc*Bc) / (Ac + Cc + X));
+    float m = sqrtf(2.0f*(Ac*Cc - 0.25f*Bc*Bc) / (Ac + Cc + X));
     
     // choose desired resolution
-    int reslog2 = PtexUtils::max(0, int(ceil(log2(0.5/m))));
+    int reslog2 = PtexUtils::max(0, int(ceilf(log2f(0.5f/m))));
 
     // convert back to triangular domain
-    A = (4/3.0) * Ac;
-    B = (2/sqrt3) * Bc + A;
-    C = -0.25 * A + 0.5 * B + Cc;
+    A = float(4/3.0) * Ac;
+    B = float(2/sqrt3) * Bc + A;
+    C = -0.25f * A + 0.5f * B + Cc;
 
     // scale by kernel width
-    double scale = PtexTriangleKernelWidth * PtexTriangleKernelWidth;
+    float scale = PtexTriangleKernelWidth * PtexTriangleKernelWidth;
     A *= scale;
     B *= scale;
     C *= scale;
 
     // find u,v,w extents
-    double uw = PtexUtils::min(sqrt(C), 1.0);
-    double vw = PtexUtils::min(sqrt(A), 1.0);
-    double ww = PtexUtils::min(sqrt(A-B+C), 1.0);
+    float uw = PtexUtils::min(sqrtf(C), 1.0f);
+    float vw = PtexUtils::min(sqrtf(A), 1.0f);
+    float ww = PtexUtils::min(sqrtf(A-B+C), 1.0f);
 
     // init kernel
-    double w = 1 - u - v;
+    float w = 1.0f - u - v;
     k.set(Res(reslog2, reslog2), u, v, u-uw, v-vw, w-ww, u+uw, v+vw, w+ww, A, B, C);
 }
